Move drop into feature to own file
